refactor favorites into storage and UI modules
